MQTT/Python/Depreciated/mqttSerialPublisher/requirements.txt
paho_mqtt==1.6.1
pyserial==3.5
